Fix additional SECURITY-333 XSS issues
Previous versions of the plugin used the <j:out> Jelly tag to write out
values, but this bypasses Jenkins's built-in XSS protections and allows
the output of raw HTML to the page. A previous commit removed many uses
of this tag to close many XSS holes, but others still remained. Remove
almost all of the remaining uses of this as it's never really needed.

One exception is with script comments, as current examples of script
comments show using HTML tags. Introduce a new dependency on the OWASP
HTML Markup Formatter and specifically pass script comments through this
formatter so they are properly sanitized, but we still allow basic HTML
formatting.

Increase the core dependency to 2.121.3 (same as the OWASP Markup
Formatter) so that we can run our tests. There are very few people still
using this old of a Jenkins version, so this should be minor bump.