Resolve possible additional XSS opportunities
The Groovy layout code used to render the Git page had some additional
XSS opportunities in it, in that it rendered code raw into the page
without escaping. Resolve this XSS vector by converting it to Jelly,
which will escape input values by default but still allow HTML in the
localization files to come through.