Simplify extension lookups
Simplify uses of Jenkins.get().getExtensionList(Class).get() to
ExtensionList.lookupSingleton(), which not only ensures that the Jenkins
singleton is non-null but also that the extension being looked up exists
and that only a single instance of it exists. An IllegalStateException
will be raised otherwise, which is simpler to handle.