Convert ScriptInfoCatalog to an interface
ScriptInfoCatalog was extending an abstract class from Ant previously to
mark itself as an extension point, which was incorrect. The correct
class is ExtensionPoint from Jenkins itself, which it was changed to
implement in a previous commit.

Given that it's only meant to define an API that catalog classes are
expected to adhere to and defines one static method for finding all
script catalogs, it makes more sense for this to be an interface.
Convert it to an interface and adapt the one implementation of it to
match.

There don't appear to be any additional plugins which depend on us and
would otherwise extend the abstract class from before, so this should be
safe. Nevertheless, mention it in the release notes.