Fix potential null dereferences
In ScriptlerManagement, ScriptHelper.getScript can potentially return
null. Check for this and set an appropriate error message if it occurs.

In ScriptlerBuilder, if we're not running on the controller node,
laucher.getChannel() can potentially return null. Check for this and
print an error if it occurs.

In ScriptlerTokenMacro, if the remote agent has disconnected, the
workspace can potentially be null. Throw an error if this occurs.