Directly deserialize Parameter instances
When receiving form data from Stapler, it comes in the form of a
JSONArray from json-lib. To use deserialization directly from that
JSONArray requires an object that has an empty constructor as well as
setter methods for each field being deserialized. This does not play
well to have an immutable object, which is better in terms of preventing
unexpected behaviour as well as simplifying use of these objects in
general (e.g., they can be used as keys for maps and sets, etc.).

In the view of wanting to make the Parameter class immutable, remove
the empty constructor in favour of one that takes a JSONObject instance
and pulls out the pertinent information from it itself. Make a new
method to loop through a JSONArray instance and map each value within
using a mapping function, which we pass our new constructor. With this,
we can remove the setters as well and make the fields final, making the
class immutable.