Convert parameter arrays to Collections
To prevent possible nulls (as some paths used null arrays rather than
empty arrays) and to promote the use of more modern classes, convert the
arrays that held Parameter instances to instances of Collection or List.
Annotate them all @NonNull and handle appropriately.

List is used in some areas where serialization is required, as XStream
will not deserialize a Collection. Collection is preferred as it's more
general and we typically don't need the ordering guarantees of a List.