JENKINS-59841: Use UTF-8 encoding to write scripts
A change between Scriptler 2.9 and 3.1 changed the logic for reading
scripts from the local disk so that it always used UTF-8 instead of the
system's default encoding, but it did not change the character set used
for writing. This can cause issues where the platform's default encoding
was not UTF-8, as we could write scripts out with a different encoding
than we read them in with. SpotBugs raised this use of the default
encoding as a warning, but apparently was missed.

Specifically use UTF-8 encoding to write scripts now so that we can now
read and write scripts always with the same encoding, regardless of
whether the system's default encoding changes or scripts are transferred
between systems with different default encodings. Add a unit test to
ensure that UTF-8 is used even if the default charset is different.