Unignore unit test by disabling CodeMirror
This one unit test was ignored supposedly because HtmlUnit didn't send
the modified textarea to the server, but the real culprit was the
CodeMirror text editor, as it "overwrote" the textarea with its own
component and then overwrote the contents of the textarea on form
submission. To bypass this, we don't apply CodeMirror to the textarea if
we're running unit tests, so the textarea contents will not be
overwritten on form submission.