Use `ArrayList` for storing parameters locally
Instead of using the Java 11-native `List.of` and `List.copyOf` to store
a copy of the parameters list locally, which aren't natively supported
by XStream serialization, use `ArrayList`, which is natively supported.
When returning the parameters, wrap them in an unmodifiable list to
avoid unexpected external modifications.

This unfortunately means that this class is not "immutable" anymore as
the parameters can technically be changed, but the current code does not
provide a way to accomplish that.