JENKINS-75019 - Use uber classloader on built-in node
Make sure that we are using the uber classloader from the plugin manager
when we're running on the built-in node. This is done by ensuring that
the `ControllerGroovyScript` class is used for running scripts when
we're running on the built-in node, which sets the right classloader.

(cherry picked from commit c46f19cb3c18840f730ccb8551a573c919cd9e54)