Fix Date formatter
An updated version of SpotBugs flags an issue with the date format
string that is used when importing a script from the script catalog.
Previously, it used a 24-hour time along with an a.m./p.m. indication,
which doesn't make much sense. Remove the a.m./p.m. indication so that
we have a proper 24-hour timestamp.